% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_clean.R
\name{tv_clean}
\alias{tv_clean}
\alias{tv_clean_names}
\alias{tv_clean_ids}
\title{Clean data}
\usage{
tv_clean_names(x, col = NULL, names = NULL, db = NULL)

tv_clean_ids(x, col = NULL, ids = NULL, db = NULL)
}
\arguments{
\item{x}{(data.frame) input data}

\item{db}{(character) database IDs from. see below for options}

\item{tax_names}{(character) column name holding taxonomic names
to use}

\item{tax_ids}{(character) column name holding taxonomic IDs
to use. if given, \code{db} is required}
}
\value{
an object of class data.frame
}
\description{
Clean data
}
\section{db options}{

\itemize{
\item \code{bold}: Barcode of Life
\item \code{col}: Catalogue of Life
\item \code{eol}: Encyclopedia of Life
\item \code{gbif}: Global Biodiversity Information Facility
\item \code{iucn}: International Union for Conservation of Nature Red List
\item \code{natserv}: Nature Serve
\item \code{nbn}: National Biodiversity Network (UK)
\item \code{tol}: Tree of Life
\item \code{tropicos}: Tropicos
\item \code{itis}: Integrated Taxonomic Information Service
\item \code{ncbi}: National Center for Biotechnology Information
\item \code{worms}: World Register of Marine Species
}
}

\examples{
\dontrun{
x <- system.file("examples/plant_spp.csv", package = "taxview")

# assuming you only have taxonomic names
# tv_clean_names(x, names = "name")

# if you have taxonomic IDs (from set of allowed databases, see above)
## if a column name
# tv_clean_ids(x, ids = "id", db = "eol")
## if a vector of IDs
dat <- tibble::as_tibble(
 data.table::fread(x, stringsAsFactors = FALSE, 
   data.table = FALSE))
out <- tv_clean_ids(x, ids = dat$id, db = "ncbi")
head(out)
}
}
