% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_summarise.R
\name{tv_summarise}
\alias{tv_summarise}
\title{Summarise data with respect to taxonomy}
\usage{
tv_summarise(x)
}
\arguments{
\item{x}{(data.frame) input data}
}
\value{
an object of class tv_summary
}
\description{
Summarise data with respect to taxonomy
}
\examples{
\dontrun{
x <- system.file("examples/plant_spp.csv", package = "taxview")
dat <- tibble::as_tibble(
 data.table::fread(x, stringsAsFactors = FALSE, 
   data.table = FALSE))
out <- tv_clean_ids(x, ids = dat$id, db = "ncbi")
(res <- tv_summarise(out))
res$summary
res$by_rank
res$by_rank_name
res$by_within_rank
}
}
