% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_viz.R
\name{tv_viz}
\alias{tv_viz}
\title{Visualize data with respect to taxonomy}
\usage{
tv_viz(x)
}
\arguments{
\item{x}{An object of class \code{tv_summary}}
}
\value{
opens interactive vis
}
\description{
Visualize data with respect to taxonomy
}
\examples{
x <- system.file("examples/plant_spp.csv", package = "taxview")
dat <- tibble::as_tibble(
 data.table::fread(x, stringsAsFactors = FALSE,
   data.table = FALSE))
out <- tv_clean_ids(x, ids = dat$id, db = "ncbi")
(res <- tv_summarise(out))
tv_viz(res)
}
